+++ Runtime/Transition.cs
@@ -7,6 +7,7 @@
 
         public Transition(IState to, IPredicate condition)
         {
+            //Implement Null Object Pattern - if IState is null, change it to the Null state object
             To = to;
             Condition = condition;
         }