+++ Samples~/Character FSM/States/CharacterMovementStates.cs
@@ -1,12 +1,27 @@
 ﻿namespace PsigenVision.Controller.Character.FiniteStateMachine
 {
-    public class CharacterStates
+	public class CharacterMovementStates
     {
+        public GroundedState Ground;
+        public FallingState Falling;
+        public SlidingState Sliding;
+        public RisingState Rising;
+        public JumpingState Jumping;
+
+        public CharacterMovementStates(ICharacterMovementCallbacks movement)
+        {
+            Ground = new GroundedState(movement);
+            Falling = new FallingState(movement);
+            Sliding = new SlidingState(movement);
+            Rising = new RisingState(movement);
+            Jumping = new JumpingState(movement);
+        }
         public class GroundedState : BaseState {
             readonly ICharacterMovementCallbacks movement;
 
             public GroundedState(ICharacterMovementCallbacks movement) {
                 this.movement = movement;
+                //Stick with BaseState.LifecycleRequirements = StateLifecycleMask.None as only Enter is used
             }
 
             public override void Enter() 
@@ -21,6 +36,7 @@
 
             public FallingState(ICharacterMovementCallbacks movement) {
                 this.movement = movement;
+                //Stick with BaseState.LifecycleRequirements = StateLifecycleMask.None as only Enter is used
             }
 
             public override void Enter() 
@@ -35,6 +51,7 @@
 
             public SlidingState(ICharacterMovementCallbacks movement) {
                 this.movement = movement;
+                //Stick with BaseState.LifecycleRequirements = StateLifecycleMask.None as only Enter is used
             }
 
             public override void Enter() 
@@ -49,6 +66,7 @@
 
             public RisingState(ICharacterMovementCallbacks movement) {
                 this.movement = movement;
+                //Stick with BaseState.LifecycleRequirements = StateLifecycleMask.None as only Enter is used
             }
 
             public override void Enter() 
@@ -63,6 +81,7 @@
 
             public JumpingState(ICharacterMovementCallbacks movement) {
                 this.movement = movement;
+                //Stick with BaseState.LifecycleRequirements = StateLifecycleMask.None as only Enter is used
             }
 
             public override void Enter()